Add set data method
