+++ src/modules/showMessage.js
@@ -1,0 +1,4 @@
+const message = (message = 'Data Inserted successfully.') => {
+  console.log(message);
+};
+export default message;
